chore:오류 수정 및 코드 간편화
