+++ src/pages/Interested/index.jsx
@@ -1,4 +1,4 @@
-import React, { useState } from "react";
+import React from "react";
 import Header from "../../duckku-ui/Header";
 import Layout from "../../duckku-ui/Layout";
 import Margin from "../../duckku-ui/Margin";
