+++ src/pages/Purchase/Modal/modal.css
@@ -2,7 +2,7 @@
   display: none;
   height: auto;
   position: fixed;
-  top: -1500px;
+  top: 0;
   right: 0;
   bottom: 0;
   left: 0;
@@ -122,7 +122,6 @@
 .modal.openModal {
   display: flex;
   align-items: center;
-
   animation: modal-bg-show 0.3s;
 }
 @keyframes modal-show {
