+++ src/pages/Wish/components/WishCard.jsx
@@ -1,4 +1,4 @@
-import React, { useState } from "react";
+import React from "react";
 import styled from "styled-components";
 import Typography from "../../../duckku-ui/Typography";
 import { AiFillHeart } from "react-icons/ai";
